+++ Changelog.md
@@ -74,3 +74,4 @@ General:
 - Agregada las fichas
 - Agregado el comando ak.buy-naturetoken
 - Agregado el comando ak.menu
+- Agregado el comando ak.cure